A little refactoring

+++ src/generate.rs
@@ -4,7 +4,7 @@ use std::collections::HashMap;
 use std::io::prelude::*;
 use std::fs::File;
 
-pub fn generate(bytes_used: u32, _variables: HashMap::<String, TypeBase>,
+pub fn generate(bytes_used: usize, _variables: HashMap::<String, TypeBase>,
                 _ast: Option<Box<Node>>, file_name: &str) -> std::io::Result<()> {
     let v: Vec<&str> = file_name.split('.').collect();
     let mut output_name = String::from(v[0]);
